requirements.txt
fastapi
uvicorn
pandas
scikit-learn
numpy
python-multipart
ydata_profiling
aiofiles
seaborn
matplotlib.pyplot
sqlalchemy
io